README.md
React boilerplate for all future projects, configured with webpack with loaders for css and babel. I might add sass support later if I find that I need that. 

If you want to try it out, just clone it, run install npm and then npm run devserver for a live-reloading server on localhost:port 3000.